tech-debt: address review findings from ROK-804 + ROK-800 batch (ROK-807)
1. Add ::int cast to DISTINCT_USER_COUNT in users-query.helpers for
   consistency with igdb-interest.helpers
2. Validate source query param against HEART_SOURCES in users.controller
3. Document rate-limit adequacy for batch pricing endpoint
4. Align getGamePricingBatch to use URLSearchParams (consistency)
5. Extract fetchGameEventTypes to igdb-event-types.helpers.ts per naming
   conventions

+++ api/src/users/users-query.helpers.ts
@@ -45,8 +45,8 @@ type UserListResult = {
   total: number;
 };
 
-/** SQL expression for distinct user count in game interest queries. */
-const DISTINCT_USER_COUNT = sql<number>`count(distinct ${schema.gameInterests.userId})`;
+/** SQL expression for distinct user count in game interest queries (::int for consistent typing with igdb-interest.helpers). */
+const DISTINCT_USER_COUNT = sql<number>`count(distinct ${schema.gameInterests.userId})::int`;
 
 /** Build WHERE conditions for game interest queries. */
 function buildGameInterestConditions(
